Further cleanup/fixes

+++ utility/common.cpp
@@ -35,13 +35,13 @@ std::u16string readNullTerminatedWStr(std::istream& in, const unsigned int offse
 }
 
 
-unsigned int padToLen(std::ostream& out, const unsigned int& len, const char pad) {
+size_t padToLen(std::ostream& out, const unsigned int& len, const char pad) {
 	if (len == 0) return 0; //don't pad to no alignment (also cant % by 0)
 
-	unsigned int padLen = len - (out.tellp() % len);
+	size_t padLen = len - (static_cast<size_t>(out.tellp()) % len);
 	if (padLen == len) return 0; //doesnt write any padding, return length 0
 
-	for (unsigned int i = 0; i < padLen; i++) {
+	for (size_t i = 0; i < padLen; i++) {
 		out.write(&pad, 1);
 	}
